fix or silence (msvc) compiler warnings about narrowing, unused formal parameters, and other minor stuff
(cherry picked from commit cc4a95bf0fc0fa71ba0cdb4f7b9d84253dc89af3)

+++ include/boost/python/suite/indexing/vector_indexing_suite.hpp
@@ -173,7 +173,7 @@ namespace boost { namespace python {
             {
                 long index = i();
                 if (index < 0)
-                    index += DerivedPolicies::size(container);
+                    index += static_cast<long>(DerivedPolicies::size(container));
                 if (index >= long(container.size()) || index < 0)
                 {
                     PyErr_SetString(PyExc_IndexError, "Index out of range");
